new component for user's details
