+++ Rentit-frontend/src/pages/login/login.jsx
@@ -1,4 +1,4 @@
-import './user.css';
+import './login.css';
 import logo from '../../images/Logo.svg';
 import { FcGoogle } from 'react-icons/fc';
 import { GoMarkGithub } from 'react-icons/go';
@@ -8,20 +8,14 @@ import Animation from '../../components/cssAnimation/animation';
 import { LoginSocialGoogle } from 'reactjs-social-login';
 import { useEffect, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
+import UserDetails from '../userDetails/userDetails';
 // import { useGlobalContext } from '../../authContext';
 
-const User = () => {
+const Login = () => {
   // const { logIn, setLogIn } = useGlobalContext();
   const navigate = useNavigate();
   const [loginData, setLoginData] = useState(null);
   const url = 'http://127.0.0.1:2000/api-rentit/v1/login';
-
-  const handleLogOut = () => {
-    sessionStorage.removeItem('userDetails');
-    // setLogIn(false);
-    // console.log(logIn);
-    navigate('/');
-  };
 
   useEffect(() => {
     const saveUser = async () => {
@@ -48,13 +42,7 @@ const User = () => {
 
   if (sessionStorage.getItem('userDetails')) {
     const data = JSON.parse(sessionStorage.getItem('userDetails'));
-    return (
-      <>
-        You are logged in
-        <div>Hi {data.given_name}</div>
-        <button onClick={handleLogOut}>Log Out</button>
-      </>
-    );
+    return <UserDetails props={data} />;
   }
 
   return (
@@ -111,4 +99,4 @@ const User = () => {
     </>
   );
 };
-export default User;
+export default Login;
